Address PR comments and fix tests
