+++ v2/codegen/apigen/maingen/testdata/multiple_services.txt
@@ -35,11 +35,12 @@ func init() {
 			Revision:    "",
 			Uncommitted: false,
 		},
-		BundledServices:   []string{"bar", "foo"},
-		CORSAllowHeaders:  nil,
-		CORSExposeHeaders: nil,
-		EncoreCompiler:    "",
-		PubsubTopics:      map[string]*config.StaticPubsubTopic{},
+		BundledServices:    []string{"bar", "foo"},
+		CORSAllowHeaders:   nil,
+		CORSExposeHeaders:  nil,
+		EnabledExperiments: nil,
+		EncoreCompiler:     "",
+		PubsubTopics:       map[string]*config.StaticPubsubTopic{},
 		TestServiceMap: map[string]string{
 			"bar": "testing_path:bar",
 			"foo": "testing_path:foo",
