+++ v2/codegen/apigen/maingen/testdata/subscription.txt
@@ -40,10 +40,11 @@ func init() {
 			Revision:    "",
 			Uncommitted: false,
 		},
-		BundledServices:   []string{"code"},
-		CORSAllowHeaders:  nil,
-		CORSExposeHeaders: nil,
-		EncoreCompiler:    "",
+		BundledServices:    []string{"code"},
+		CORSAllowHeaders:   nil,
+		CORSExposeHeaders:  nil,
+		EnabledExperiments: nil,
+		EncoreCompiler:     "",
 		PubsubTopics: map[string]*config.StaticPubsubTopic{"topic": {Subscriptions: map[string]*config.StaticPubsubSubscription{"subscription": {
 			Service:  "code",
 			SvcNum:   1,
